fix(critic): clear stale outfile, int line, non-masking fence restore, schema-path guard (codex gate)

+++ src/critic/verdict.ts
@@ -25,7 +25,7 @@ const BrokenContractSchema = z
   .object({
     zone: z.string(),
     file: z.string(),
-    line: z.number(),
+    line: z.number().int(),
     evidence: z.string(),
   })
   .strict();
